refactor(cli): abstract command class
so, the cobra approach doesn't work in pure form, because there is
essentially no global namespace to create commands from

and just do the commands in the cli class is invalid as well

so my solution is to just make the command class abstract to implement
it. it's kinda awkward because i need the validation for the usage
string, but it came out good

the checks via the runners class is not so good either, but it works
fine. also there is a bit of boilerplate, but whatever